Move AStarPathFinder to shared library
