+++ EndlessClient/Rendering/Character/CharacterAnimator.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using EndlessClient.GameExecution;
-using EndlessClient.HUD;
 using EndlessClient.HUD.Spells;
 using EndlessClient.Rendering.Metadata;
 using EndlessClient.Rendering.Metadata.Models;
@@ -12,6 +11,7 @@ using EOLib.Domain.Character;
 using EOLib.Domain.Extensions;
 using EOLib.Domain.Login;
 using EOLib.Domain.Map;
+using EOLib.Domain.Pathing;
 using EOLib.Domain.Spells;
 using EOLib.IO.Map;
 using EOLib.IO.Pub;
